pyproject.toml
[project]
name = "aimc"
version = "0.1.0"
description = "AIMC – AI-assisted project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "mistralai>=1.12.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "stormvogel>=0.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aimc"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "black>=26.1.0",
    "httpx>=0.28.1",
]
